+++ python/tank/descriptor/descriptor_installed_config.py
@@ -12,6 +12,7 @@ from __future__ import with_statement
 
 import os
 
+from .descriptor import Descriptor
 from .descriptor_config import ConfigDescriptor
 from .. import pipelineconfig_utils
 from .. import LogManager
@@ -32,17 +33,25 @@ class InstalledConfigDescriptor(ConfigDescriptor):
     environment files, a copy of core and all the bundles required by that pipeline configuration.
     It supports localized as well as shared core and as such, the interpreter files can be found
     inside the configuration folder or alongside the shared core.
-
-    :param sg_connection: Connection to the Shotgun site.
-    :type sg_connection: :class:`shotgun_api3.Shotgun`
-    :param str bundle_cache_root: Path to the global bundle cache.
-    :param list(str) fallback_roots: List of fallback locations on disk where bundles can be found.
-    :param io_descriptor: Associated IO descriptor.
     """
 
-    def __init__(self, sg_connection, bundle_cache_root, fallback_roots, io_descriptor):
+    def __init__(self, sg_connection, io_descriptor, bundle_cache_root_override, fallback_roots):
+        """
+        .. note:: Use the factory method :meth:`create_descriptor` when
+                  creating new descriptor objects.
+
+        :param sg_connection: Connection to the current site.
+        :param io_descriptor: Associated IO descriptor.
+        :param bundle_cache_root_override: Override for root path to where
+            downloaded apps are cached.
+        :param fallback_roots: List of immutable fallback cache locations where
+            apps will be searched for.
+        """
         super(InstalledConfigDescriptor, self).__init__(
-            sg_connection, bundle_cache_root, fallback_roots, io_descriptor
+            sg_connection,
+            io_descriptor,
+            bundle_cache_root_override,
+            fallback_roots
         )
         self._io_descriptor.set_is_copiable(False)
 
@@ -193,3 +202,9 @@ class InstalledConfigDescriptor(ConfigDescriptor):
         return ShotgunPath.get_file_name_from_template(
             os.path.join(install_root, "install", "core", "core_%s.cfg")
         )
+
+
+# register the descriptor subclass with the base class factory
+# this complex process for handling the descriptor abstract factory
+# management is in order to avoid local imports in classes.
+Descriptor.register_descriptor_factory(Descriptor.INSTALLED_CONFIG, InstalledConfigDescriptor)
