+++ tests/descriptor_tests/test_descriptors.py
@@ -40,7 +40,15 @@ class TestCachedConfigDescriptor(ShotgunTestBase):
             def resolve_core_descriptor(self):
                 io_desc = Mock()
                 io_desc.get_manifest.return_value = dict()
-                return CoreDescriptor(io_desc)
+                sg_connection = Mock()
+                bundle_cache_root_override = None
+                fallback_roots = None
+                return CoreDescriptor(
+                    sg_connection,
+                    io_desc,
+                    bundle_cache_root_override,
+                    fallback_roots
+                )
 
         desc = CoreConfigDescriptorWithoutFeatures(None, None, None, None)
         self.assertIsNone(
@@ -55,7 +63,15 @@ class TestCachedConfigDescriptor(ShotgunTestBase):
             def resolve_core_descriptor(self):
                 io_desc = Mock()
                 io_desc.get_manifest.return_value = dict(features=dict(two="2"))
-                return CoreDescriptor(io_desc)
+                sg_connection = Mock()
+                bundle_cache_root_override = None
+                fallback_roots = None
+                return CoreDescriptor(
+                    sg_connection,
+                    io_desc,
+                    bundle_cache_root_override,
+                    fallback_roots
+                )
 
         desc = CoreConfigDescriptorWithFeatures(None, None, None, None)
 
@@ -823,13 +839,27 @@ class TestConstraintValidation(unittest2.TestCase):
 
 class TestFeaturesApi(unittest2.TestCase):
 
+    def _create_core_desc(self, io_descriptor):
+        """
+        Helper method which creates an io_descriptor
+        """
+        sg_connection = Mock()
+        bundle_cache_root_override = None
+        fallback_roots = None
+        return sgtk.descriptor.CoreDescriptor(
+            sg_connection,
+            io_descriptor,
+            bundle_cache_root_override,
+            fallback_roots
+        )
+
     def test_missing_manifest(self):
         """
         Ensures a missing manifest is handled properly.
         """
         io_desc = Mock()
         io_desc.get_manifest.side_effect = TankMissingManifestError()
-        desc = sgtk.descriptor.CoreDescriptor(io_desc)
+        desc = self._create_core_desc(io_desc)
 
         self.assertEqual(desc.get_feature_info("missing", "value"), "value")
         self.assertIsNone(desc.get_feature_info("missing"))
@@ -841,7 +871,7 @@ class TestFeaturesApi(unittest2.TestCase):
         """
         io_desc = Mock()
         io_desc.get_manifest.return_value = {}
-        desc = sgtk.descriptor.CoreDescriptor(io_desc)
+        desc = self._create_core_desc(io_desc)
 
         self.assertEqual(desc.get_feature_info("missing", "value"), "value")
         self.assertIsNone(desc.get_feature_info("missing"))
@@ -853,7 +883,7 @@ class TestFeaturesApi(unittest2.TestCase):
         """
         io_desc = Mock()
         io_desc.get_manifest.return_value = dict(features={})
-        desc = sgtk.descriptor.CoreDescriptor(io_desc)
+        desc = self._create_core_desc(io_desc)
 
         self.assertEqual(desc.get_feature_info("missing", "value"), "value")
         self.assertIsNone(desc.get_feature_info("missing"))
@@ -866,7 +896,7 @@ class TestFeaturesApi(unittest2.TestCase):
         features = dict(two="2", foo="bar", zero=0)
         io_desc = Mock()
         io_desc.get_manifest.return_value = dict(features=features)
-        desc = sgtk.descriptor.CoreDescriptor(io_desc)
+        desc = self._create_core_desc(io_desc)
 
         self.assertEqual(desc.get_feature_info("two", 3), "2")
         self.assertEqual(desc.get_feature_info("two"), "2")
@@ -896,7 +926,7 @@ class TestFeaturesApi(unittest2.TestCase):
 
         io_desc = Mock()
         io_desc.get_manifest.return_value = info
-        desc = sgtk.descriptor.CoreDescriptor(io_desc)
+        desc = self._create_core_desc(io_desc)
 
         features = {
             "bootstrap.lean_config.version": 1
